refactor(index): extract the file-preparation pipeline to prep.rs
Move the parse/chunk/symbol-into-prepared-rows pipeline (collect_index_files,
collect_changed_index_files, prepare_chunks, prepare_index_content,
prepare_files_with_progress, git-history prepare spawn/join, wave sizing) + the
Prepared* output types out of index/mod.rs into prep.rs. pub(crate) + re-export;
Prepared* fields pub(crate) (read by file_index during insert). mod.rs 914 -> 622.
No behavior change.
