+++ crates/rag-rat-core/src/index/mod.rs
@@ -23,6 +23,7 @@ mod lifecycle;
 mod meta;
 mod packages;
 mod parser_failures;
+mod prep;
 mod query_api;
 mod rebuild;
 mod staleness;
@@ -31,6 +32,7 @@ pub(crate) use discovery::*;
 pub use git_context::resolve_git_context;
 pub(crate) use git_context::*;
 pub(crate) use lifecycle::install_scope_view;
+pub(crate) use prep::*;
 pub use query_api::ImportantSymbolsRequest;
 pub(crate) use util::*;
 
@@ -553,160 +555,6 @@ impl FileScope {
     }
 }
 
-#[derive(Debug)]
-struct PreparedIndexFile {
-    file: IndexFile,
-    prepared: anyhow::Result<PreparedIndexContent>,
-}
-
-/// A chunk with all of its per-chunk CPU work (text hash, relocation anchor, embedding policy)
-/// precomputed in the parallel prepare phase. The serial insert stage then only steps the INSERT —
-/// no hashing on the single-threaded path.
-#[derive(Debug)]
-struct PreparedChunk {
-    chunk: Chunk,
-    text_hash: String,
-    anchor: anchors::ChunkAnchor,
-    embedding: ai::EmbeddingPolicyDecision,
-}
-
-/// Compute the per-chunk hashes / anchor / embedding policy for a file's chunks. Splits the file's
-/// lines once (anchors only read a few boundary/context lines). Shared by the parallel prepare
-/// phase and the inline incremental path, so both keep this CPU work off the serial insert stage.
-fn prepare_chunks(
-    path: &Path,
-    language: &str,
-    file_kind: &str,
-    chunks: Vec<Chunk>,
-    full_text: &str,
-) -> Vec<PreparedChunk> {
-    let full_lines = full_text.lines().collect::<Vec<_>>();
-    chunks
-        .into_iter()
-        .map(|chunk| {
-            let text_hash = hex_sha256(chunk.text.as_bytes());
-            let anchor = anchors::anchor_for_lines(
-                &chunk.text,
-                chunk.start_line,
-                chunk.end_line,
-                &full_lines,
-            );
-            let embedding = ai::embedding_policy_for_chunk(
-                path,
-                language,
-                file_kind,
-                chunk.kind,
-                chunk.symbol_path.as_deref(),
-                &chunk.text,
-                ai::DEFAULT_MAX_EMBEDDING_CHARS,
-            );
-            PreparedChunk { chunk, text_hash, anchor, embedding }
-        })
-        .collect()
-}
-
-#[derive(Debug)]
-struct PreparedIndexContent {
-    modified_at_ms: i64,
-    sha256: String,
-    chunks: Vec<PreparedChunk>,
-    symbols: Vec<Symbol>,
-    // Graph edge candidates computed here in the parallel prepare phase (their `from_symbol_id`s
-    // are local symbol indices, remapped to DB ids in insert_prepared_file). Empty for
-    // generated / oversized / markdown files. Moving this off the serial insert loop kills the
-    // duplicate parse.
-    edge_candidates: Vec<edges::EdgeCandidate>,
-    parser_failure: Option<String>,
-}
-
-fn collect_index_files(config: &Config) -> anyhow::Result<Vec<IndexFile>> {
-    let mut targets = config.targets.iter().collect::<Vec<_>>();
-    targets.sort_by_key(|target| match target.kind {
-        TargetKind::Generated => 0,
-        TargetKind::Tests => 1,
-        TargetKind::Docs => 2,
-        TargetKind::Source => 3,
-    });
-    let mut seen = BTreeSet::new();
-    let mut files = Vec::new();
-
-    // Compile the repo's `.gitignore` rules (ancestor chain + nested-along-target-trees) once,
-    // shared across targets, so the walker skips the same paths the watcher's `event_is_relevant`
-    // will (issue #62). Nested-gitignore discovery is scoped to the configured target directories
-    // so a large unindexed sibling is never walked (round-3 finding).
-    let ignore = ignore_rules::IgnoreMatcher::compile(&config.root, &config.target_directories());
-    for target in targets {
-        for file in walker::walk_target(&config.root, target, &ignore)? {
-            let relative_path = file.strip_prefix(&config.root)?.to_path_buf();
-            if !seen.insert(relative_path.clone()) {
-                continue;
-            }
-            files.push(IndexFile {
-                full_path: file,
-                relative_path,
-                language: target.language,
-                kind: target.kind,
-                commit_sha: String::new(),
-                worktree_id: String::new(),
-            });
-        }
-    }
-
-    Ok(files)
-}
-
-fn collect_changed_index_files(
-    config: &Config,
-    changes: &GitChangedPaths,
-) -> anyhow::Result<Vec<IndexFile>> {
-    let mut files = Vec::new();
-    for relative_path in &changes.changed {
-        let full_path = config.root.join(relative_path);
-        if !full_path.is_file() {
-            continue;
-        }
-        let Some((language, kind)) = target_for_path(config, relative_path) else {
-            continue;
-        };
-        files.push(IndexFile {
-            full_path,
-            relative_path: relative_path.clone(),
-            language,
-            kind,
-            commit_sha: String::new(),
-            worktree_id: String::new(),
-        });
-    }
-    Ok(files)
-}
-
-fn spawn_git_history_prepare(
-    root: &Path,
-) -> JoinHandle<anyhow::Result<git_history::PreparedGitHistory>> {
-    let root = root.to_path_buf();
-    thread::spawn(move || git_history::prepare(&root))
-}
-
-fn join_git_history_prepare(
-    handle: JoinHandle<anyhow::Result<git_history::PreparedGitHistory>>,
-) -> anyhow::Result<git_history::PreparedGitHistory> {
-    handle.join().map_err(|_| anyhow::anyhow!("git history preparation panicked"))?
-}
-
-fn prepare_index_file(file: &IndexFile) -> PreparedIndexFile {
-    PreparedIndexFile { file: file.clone(), prepared: prepare_index_content(file) }
-}
-
-/// Files per full-rebuild wave (prepare → insert → drop). Bounds peak memory; tunable via
-/// `RAG_RAT_INDEX_WAVE` for memory/throughput trade-offs on a given machine.
-fn index_wave_size() -> usize {
-    std::env::var("RAG_RAT_INDEX_WAVE")
-        .ok()
-        .and_then(|value| value.parse::<usize>().ok())
-        .filter(|&size| size > 0)
-        .unwrap_or(2_000)
-}
-
 /// Diagnostic: when `RAG_RAT_SQLITE_SOFT_HEAP_LIMIT_MB` is set, cap SQLite's soft heap limit
 /// (process-wide) for this run. Distinguishes a *discretionary* SQLite buffer spike (clamps with
 /// little wall cost — the buffer just flushes earlier) from a *load-bearing* allocation (wall
@@ -768,147 +616,6 @@ pub(crate) fn mem_trace(label: &str) {
         sqlite_bytes as f64 / 1_073_741_824.0,
         sqlite_peak as f64 / 1_073_741_824.0,
     );
-}
-
-/// Prepare a slice of files in parallel. `base` / `grand_total` let a caller that processes files
-/// in waves report progress against the overall total rather than the wave size (a non-waved caller
-/// passes `0` and `files.len()`).
-fn prepare_files_with_progress<F>(
-    files: &[IndexFile],
-    progress: &mut F,
-    base: usize,
-    grand_total: usize,
-) -> anyhow::Result<Vec<PreparedIndexFile>>
-where
-    F: FnMut(IndexProgress),
-{
-    #[derive(Debug)]
-    struct PreparedProgress {
-        current: usize,
-        total: usize,
-        path: PathBuf,
-        language: Language,
-        kind: TargetKind,
-    }
-
-    let prepared = thread::scope(|scope| {
-        let (tx, rx) = mpsc::channel();
-        let completed = AtomicUsize::new(0);
-        let handle = scope.spawn(move || {
-            files
-                .par_iter()
-                .map(|file| {
-                    let prepared = prepare_index_file(file);
-                    let current = base + completed.fetch_add(1, Ordering::Relaxed) + 1;
-                    if should_report_file_progress(current, grand_total) {
-                        let _ = tx.send(PreparedProgress {
-                            current,
-                            total: grand_total,
-                            path: file.relative_path.clone(),
-                            language: file.language,
-                            kind: file.kind,
-                        });
-                    }
-                    prepared
-                })
-                .collect::<Vec<_>>()
-        });
-
-        for event in rx {
-            progress(IndexProgress::PreparingFile {
-                current: event.current,
-                total: event.total,
-                path: event.path,
-                language: event.language,
-                kind: event.kind,
-            });
-        }
-
-        handle.join().map_err(|_| anyhow::anyhow!("parallel file preparation panicked"))
-    })?;
-    Ok(prepared)
-}
-
-fn should_report_file_progress(current: usize, total: usize) -> bool {
-    if total == 0 {
-        return false;
-    }
-    current == 1
-        || current == total
-        || current.saturating_mul(10) / total
-            != current.saturating_sub(1).saturating_mul(10) / total
-}
-
-fn prepare_index_content(file: &IndexFile) -> anyhow::Result<PreparedIndexContent> {
-    let text = fs::read_to_string(&file.full_path)?;
-    let modified_at_ms = file_metadata_ms(&file.full_path)?;
-    let sha256 = hex_sha256(text.as_bytes());
-
-    // ONE tree-sitter parse per file, shared by chunks + symbols + edges + the parser-failure flag,
-    // instead of re-parsing the same file four times. Only for structural (non-generated,
-    // non-markdown) code within the parse-size cap; everything else takes the line-based paths.
-    let structural_eligible = file.kind != TargetKind::Generated
-        && file.language != Language::Markdown
-        && text.len() <= chunker::MAX_STRUCTURAL_PARSE_BYTES;
-    let parsed = structural_eligible
-        .then(|| parser::parse_file(&file.relative_path, file.language, &text))
-        .flatten();
-
-    let symbols = parsed.as_ref().map(|p| symbols::from_parsed(&p.symbols)).unwrap_or_default();
-
-    // Preserve the historical failure signal: a clean parse → None, error nodes → the message, and
-    // a hard parse failure on otherwise-eligible code → an error (matches the old parse_error
-    // Err arm).
-    let parser_failure = if structural_eligible {
-        match &parsed {
-            Some(p) => p.parser_failure(),
-            None => Some("tree-sitter parse failed".to_string()),
-        }
-    } else {
-        None
-    };
-
-    let chunks = if file.kind == TargetKind::Generated {
-        chunker::generated_chunks_for_file(&file.relative_path, &text)
-    } else if let Some(p) = &parsed {
-        chunker::code_chunks_for_symbols(&file.relative_path, &text, &p.symbols)
-    } else {
-        // Markdown, oversized, or a hard parse failure: line-based chunking, no shared tree.
-        chunker::chunks_for_file(&file.relative_path, file.language, &text)
-    };
-    // Precompute per-chunk hashes / anchor / embedding policy here, in parallel.
-    let chunks = prepare_chunks(
-        &file.relative_path,
-        file.language.as_str(),
-        file.kind.as_str(),
-        chunks,
-        &text,
-    );
-
-    // Edge candidates walk the shared tree (no re-parse). from_symbol_id holds a local symbol
-    // index, remapped to the real DB id at insert time. Empty when there's no structural parse.
-    let edge_candidates = match &parsed {
-        Some(p) => {
-            let local = edges::IndexedSymbol::local_from_prepared(file.language, &symbols);
-            edges::edge_candidates_from_root(
-                &file.relative_path,
-                file.language,
-                &text,
-                p.root(),
-                &local,
-            )
-        },
-        None => Vec::new(),
-    };
-
-    Ok(PreparedIndexContent {
-        modified_at_ms,
-        sha256,
-        chunks,
-        symbols,
-        edge_candidates,
-        parser_failure,
-    })
 }
 
 #[cfg(test)]
